bytepad/requirements.txt
Django==1.10.6
django-watson==1.3.0
djangorestframework==3.6.2
psycopg2==2.7.1
